Make a quick fix for `behindIs(phrase(<keyword>))` situations
Because `phrase(...)` translates to `concat(behindIs([^a-z]), ..., aheadIs([^a-z]))`, and because the root `behindIs()` will splice out the `aheadIs()`, this leads to situations where `behindIs(phrase(...))` will feel like it's acting weird. For example, `(const)` in `(const) myThing` will be accepted by this pattern (at the ` `) when it feels like it shouldn't. It feels this way because ` ` *should* be the character recognized by the original `aheadIs()`.

Since I'm not sure how to fix this yet (or if it's an actual problem with `phrase()` or `behindIs()`), I added some TODOs for the hacks I made